requirements.txt
openai==1.36.0
PyMuPDF==1.24.9
python-telegram-bot==21.4
python-dotenv
pillow==10.2.0
sglang==0.2.13
pydantic==2.8.2
outlines==0.0.46